refactor chord-practice page
